swissairdry-installation/docker-base-api/app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.19
pydantic>=2.3.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.2
psycopg2-binary>=2.9.7
bcrypt>=4.0.1
pandas>=2.1.0
aiofiles>=23.2.1
requests>=2.31.0
httpx>=0.24.1
markdown>=3.4.4
pillow>=10.0.1
paho-mqtt>=2.2.1